skin-cancer-api/python_model/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
numpy
tensorflow